+++ src/Commands/LaravelEnvcoderDecrypt.php
@@ -6,14 +6,13 @@ use harmonic\LaravelEnvcoder\LaravelEnvcoder;
 use Illuminate\Contracts\Filesystem\FileNotFoundException;
 use harmonic\LaravelEnvcoder\Facades\LaravelEnvcoder as LEFacade;
 
-class LaravelEnvcoderDecrypt extends \harmonic\LaravelEnvcoder\LaravelEnvcoderBaseCommand
-{
+class LaravelEnvcoderDecrypt extends \harmonic\LaravelEnvcoder\LaravelEnvcoderBaseCommand {
     /**
      * The name and signature of the console command.
      *
      * @var string
      */
-    protected $signature = 'env:decrypt {--password=}';
+    protected $signature = 'env:decrypt {--p|password=} {--s|source=.env.enc}';
 
     /**
      * The console command description.
@@ -27,8 +26,7 @@ class LaravelEnvcoderDecrypt extends \harmonic\LaravelEnvcoder\LaravelEnvcoderBa
      *
      * @return void
      */
-    public function __construct()
-    {
+    public function __construct() {
         parent::__construct();
     }
 
@@ -37,14 +35,14 @@ class LaravelEnvcoderDecrypt extends \harmonic\LaravelEnvcoder\LaravelEnvcoderBa
      *
      * @return mixed
      */
-    public function handle()
-    {
+    public function handle() {
         $key = $this->getPassword();
         $envcoder = new LaravelEnvcoder();
         $resolve = config('laravel-envcoder.resolve');
+        $sourceEnv = $this->option('source');
 
         try {
-            $result = $envcoder->decrypt($key);
+            $result = $envcoder->decrypt($key, $sourceEnv);
 
             if ($resolve === 'merge') {
                 if ($result) {
@@ -55,28 +53,28 @@ class LaravelEnvcoderDecrypt extends \harmonic\LaravelEnvcoder\LaravelEnvcoderBa
                     $envFile = fopen('.env', 'w');
                     foreach ($result['decrypted'] as $key => $value) {
                         if (array_key_exists($key, $result['current']) && $value !== $result['current'][$key]) {
-                            $use = $this->choice('Env variable '.$key.' has encrypted value (E) '.$value.' vs unencrypted value (U) '.$result['current'][$key], ['E', 'U'], 0);
+                            $use = $this->choice('Env variable ' . $key . ' has encrypted value (E) ' . $value . ' vs unencrypted value (U) ' . $result['current'][$key], ['E', 'U'], 0);
                             if ($use === 'E') {
                                 $value = LEFacade::formatValue($value);
-                                fwrite($envFile, $key.'='.$value.PHP_EOL);
+                                fwrite($envFile, $key . '=' . $value . PHP_EOL);
                                 continue;
                             }
-                        } elseif (! array_key_exists($key, $result['current'])) {
-                            $use = $this->choice('Env variable '.$key.' has encrypted value '.$value.' but does not exist in .env add (A) or skip (S)', ['A', 'S'], 0);
+                        } elseif (!array_key_exists($key, $result['current'])) {
+                            $use = $this->choice('Env variable ' . $key . ' has encrypted value ' . $value . ' but does not exist in .env add (A) or skip (S)', ['A', 'S'], 0);
                             if ($use === 'S') {
                                 continue;
                             }
                         }
-                        fwrite($envFile, $key.'='.$result['decrypted'][$key].PHP_EOL);
+                        fwrite($envFile, $key . '=' . $result['decrypted'][$key] . PHP_EOL);
                     }
                     $varsNotYetAdded = array_diff_key($result['current'], $result['decrypted']);
                     foreach ($varsNotYetAdded as $key => $value) {
-                        $use = $this->choice('Env variable '.$key.' with value '.$value.' found in .env not in .env.enc add (A) or skip (S)', ['A', 'S'], 0);
+                        $use = $this->choice('Env variable ' . $key . ' with value ' . $value . ' found in .env not in .env.enc add (A) or skip (S)', ['A', 'S'], 0);
                         if ($use === 'S') {
                             continue;
                         }
                         $value = LEFacade::formatValue($value);
-                        fwrite($envFile, $key.'='.$value.PHP_EOL);
+                        fwrite($envFile, $key . '=' . $value . PHP_EOL);
                     }
 
                     fclose($envFile);