+++ src/LaravelEnvcoder.php
@@ -6,15 +6,13 @@ use Defuse\Crypto\File;
 use Illuminate\Contracts\Filesystem\FileNotFoundException;
 use harmonic\LaravelEnvcoder\Facades\LaravelEnvcoder as LEFacade;
 
-class LaravelEnvcoder
-{
+class LaravelEnvcoder {
     /**
      * Check in the ENV_PASSWORD .env variable exists and return it if so.
      *
      * @return string if exists, null otherwise
      */
-    public function getPasswordFromEnv() : ?string
-    {
+    public function getPasswordFromEnv() : ?string {
         $envArray = $this->envToArray('.env');
         if (array_key_exists('ENV_PASSWORD', $envArray)) {
             return $envArray['ENV_PASSWORD'];
@@ -27,18 +25,18 @@ class LaravelEnvcoder
      * Encrypts the .env file and saves as .env.enc.
      *
      * @param string $password The password used to encrypt the file
+     * @param string $sourceEnv The name of the .env file to encrypt
      * @return void
      */
-    public function encrypt(string $password) : void
-    {
+    public function encrypt(string $password, string $sourceEnv) : void {
         $needsPasswordAdded = false;
         if ($this->getPasswordFromEnv() !== null) {
-            $contents = file_get_contents('.env');
-            $contents = preg_replace('/ENV_PASSWORD='.$password.'\R?/', '', $contents);
-            file_put_contents('.env', $contents);
+            $contents = file_get_contents($sourceEnv);
+            $contents = preg_replace('/ENV_PASSWORD=' . $password . '\R?/', '', $contents);
+            file_put_contents($sourceEnv, $contents);
             $needsPasswordAdded = true;
         }
-        File::encryptFileWithPassword('.env', '.env.enc', $password);
+        File::encryptFileWithPassword($sourceEnv, $sourceEnv . '.enc', $password);
         if ($needsPasswordAdded) {
             $this->addPasswordToEnv($password);
         }
@@ -51,10 +49,9 @@ class LaravelEnvcoder
      * @param string $password
      * @return void
      */
-    private function addPasswordToEnv(string $password)
-    {
+    private function addPasswordToEnv(string $password) {
         $handle = fopen('.env', 'a');
-        fwrite($handle, 'ENV_PASSWORD='.$password.PHP_EOL);
+        fwrite($handle, 'ENV_PASSWORD=' . $password . PHP_EOL);
         fclose($handle);
     }
 
@@ -64,9 +61,8 @@ class LaravelEnvcoder
      * @param string $env
      * @return array
      */
-    public function envToArray(string $envFile) : array
-    {
-        if (! is_file($envFile)) {
+    public function envToArray(string $envFile) : array {
+        if (!is_file($envFile)) {
             return [];
         }
 
@@ -77,11 +73,11 @@ class LaravelEnvcoder
      * Decrypt the .env.enc file using.
      *
      * @param string $password The password used to encrypt the file
+     * @param string $sourceEnv The name of the encrypted .env file to decrypt
      * @return void
      */
-    public function decrypt(string $password)
-    {
-        if (! \file_exists('.env.enc')) {
+    public function decrypt(string $password, string $sourceEnv) {
+        if (!\file_exists($sourceEnv)) {
             throw new FileNotFoundException('No encrypted env file found.');
         }
 
@@ -92,22 +88,25 @@ class LaravelEnvcoder
             $needsPasswordAdded = true;
         }
 
+        $sourceEnvUnencrypted = substr($sourceEnv, 0, -4);
+
         switch ($resolve) {
             case 'ignore':
                 return;
             case 'overwrite':
-                File::decryptFileWithPassword('.env.enc', '.env', $password);
+                File::decryptFileWithPassword($sourceEnv, $sourceEnvUnencrypted, $password);
                 if ($needsPasswordAdded) {
                     $this->addPasswordToEnv($password);
                 }
 
                 return;
             case 'prompt':
-                File::decryptFileWithPassword('.env.enc', '.env.bak', $password);
-                $decryptedArray = $this->envToArray('.env.bak');
-                $currentEnv = $this->envToArray('.env');
+                $envBak = $sourceEnvUnencrypted . '.bak';
+                File::decryptFileWithPassword($sourceEnv, $envBak, $password);
+                $decryptedArray = $this->envToArray($envBak);
+                $currentEnv = $this->envToArray($sourceEnvUnencrypted);
 
-                unlink('.env.bak');
+                unlink($envBak);
 
                 if ($needsPasswordAdded) {
                     $currentEnv['ENV_PASSWORD'] = $password;
@@ -118,20 +117,21 @@ class LaravelEnvcoder
                     'current' => $currentEnv,
                 ];
             case 'merge':
-                File::decryptFileWithPassword('.env.enc', '.env.bak', $password);
-                $decryptedArray = $this->envToArray('.env.bak');
+                $envBak = $sourceEnvUnencrypted . '.bak';
+                File::decryptFileWithPassword($sourceEnv, $envBak, $password);
+                $decryptedArray = $this->envToArray($envBak);
                 if ($needsPasswordAdded) {
                     $decryptedArray['ENV_PASSWORD'] = $password;
                 }
-                $currentEnv = $this->envToArray('.env');
+                $currentEnv = $this->envToArray($sourceEnvUnencrypted);
                 $mergedArray = array_merge($currentEnv, $decryptedArray);
-                $envFile = fopen('.env', 'w');
+                $envFile = fopen($sourceEnvUnencrypted, 'w');
                 foreach ($mergedArray as $key => $value) {
                     $value = LEFacade::formatValue($value);
-                    fwrite($envFile, $key.'='.$value.PHP_EOL);
+                    fwrite($envFile, $key . '=' . $value . PHP_EOL);
                 }
                 fclose($envFile);
-                unlink('.env.bak');
+                unlink($envBak);
 
                 if (count($mergedArray) > count($decryptedArray)) {
                     return true; // let the calling function know something was merged and they potentially need to encrypt to sync
@@ -152,9 +152,8 @@ class LaravelEnvcoder
      * @throws WrongKeyOrModifiedCiphertextException
      * @throws FileNotFoundException
      */
-    public function compare(string $password) : array
-    {
-        if (! \file_exists('.env.enc')) {
+    public function compare(string $password) : array {
+        if (!\file_exists('.env.enc')) {
             throw new FileNotFoundException('No encrypted env file found.');
         }